lobs/investment_platform/requirements.txt
# Microsoft Agent Framework (pinned for stability)
agent-framework-azure-ai==1.0.0b260107
agent-framework-core==1.0.0b260107
opentelemetry-semantic-conventions-ai==0.4.13  # v0.5.0 breaks agent-framework (removed LLM_* attrs)
 
# FastAPI and Web Server
fastapi==0.135.1
uvicorn[standard]==0.42.0
pydantic==2.12.5
 
# Agent HTTP Server support (legacy - keeping for reference)
azure-ai-agentserver-core==1.0.0b10
azure-ai-agentserver-agentframework==1.0.0b10
azure-ai-projects==2.0.0b3  # 2.0.0+ removed PromptAgentDefinitionText needed by agent-framework b260107
 
# Debugging support
debugpy==1.8.20
agent-dev-cli==0.0.1b260205
 
# Azure authentication
azure-identity==1.25.3
 
# GitHub integration
PyGithub==2.8.1
 
# Environment variables
python-dotenv==1.2.2
 
# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
aiohttp==3.13.3
 
# Terminal colors
colorama==0.4.6
 
 